[PHP] Share fcntl locking syscall adapter
